https://github.com/rytilahti/python-miio/pull/1174#pullrequestreview-799144026
